backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlmodel>=0.0.16
psycopg2-binary>=2.9
python-dotenv>=1.0
alembic>=1.13

